:boom: Move styles to src (#50)
