+++ frontend/src/pages/_app.tsx
@@ -1,6 +1,6 @@
 import { AppProps } from 'next/app';
 
-import '../../styles/globals.css';
+import '../styles/globals.css';
 
 /* eslint-disable react/jsx-props-no-spreading */
 function MyApp({ Component, pageProps }: AppProps): JSX.Element {
